requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://docs.astronomer.io/astro/runtime-image-architecture#provider-packages
pydantic==2.8.2
requests==2.32.3
pandas
google-cloud-storage==2.17.0
apache-airflow-providers-google==10.21.0
argparse
apache-airflow-providers-dbt-cloud